requirements.txt
Flask
Flask-static
gunicorn
azure-cosmos
jsonschema
datetime
tqdm
